pyproject.toml
[tool.poetry]
name = "gpspy"
version = "0.1.0"
description = "GPS Decoder Frontend"
authors = ["Patrick Dreker <patrick@dreker.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"
pynmea2 = "^1.18.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
flake8-bandit = "^4.1.1"
Flake8-pyproject = "^1.2.2"
mypy = "^0.991"
flake8-bugbear = "^22.12.6"
flake8-comprehensions = "^3.10.1"
flake8-requirements = "^1.7.5"
flake8-warnings = "^0.4.0"
flake8-pie = "^0.16.0"
pep8-naming = "^0.13.3"
flake8-isort = "^6.0.0"
flake8-pylint = "^0.1.3"
flake8-black = "^0.3.6"

[tool.flake8]
max-line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
